move fakeRtmClient into its own file
