+++ slack/bot_test.go
@@ -4,35 +4,7 @@ import (
 	"log"
 	"os"
 	"testing"
-
-	"github.com/nlopes/slack"
 )
-
-type fakeRtmClient struct {
-	GetUsersInConversationHandler func(*slack.GetUsersInConversationParameters) ([]string, string, error)
-	DisconnectHandler             func() error
-	PostMessageHandler            func(string, string, slack.PostMessageParameters) (string, string, error)
-}
-
-func (f *fakeRtmClient) GetUsersInConversation(params *slack.GetUsersInConversationParameters) ([]string, string, error) {
-	return f.GetUsersInConversationHandler(params)
-}
-
-func (f *fakeRtmClient) GetInfo() *slack.Info {
-	return &slack.Info{User: &slack.UserDetails{ID: "Caratroc"}}
-}
-
-func (f *fakeRtmClient) GetUserInfo(user string) (*slack.User, error) {
-	return &slack.User{ID: "84", Name: "Caratroc"}, nil
-}
-
-func (f *fakeRtmClient) PostMessage(channel, text string, params slack.PostMessageParameters) (string, string, error) {
-	return f.PostMessageHandler(channel, text, params)
-}
-
-func (f *fakeRtmClient) Disconnect() error {
-	return f.DisconnectHandler()
-}
 
 func TestHandleHaltCallsDisconnectOnClient(t *testing.T) {
 	disconnectCalled := false
